Visual Demonstration of Deep Learning Algorithm
This is my deep learning algorithm visualized. To see actual node values, please refer to the table. The values {8192, 128, 128, 8192, 64, 128, 128, 192, 64} were divided by 64 (GCD) to produce {2304, 512, 128, 2, 2, 128, 1, 2, 2, 3, 1}. I could not render the node values of {147456, 32768}, so I spaced them out to differentiate them from the other inner layers. Also, input and outer layers are not uniform in terms of size to hidden layers (otherwise, their scale would be 1/64 and 3/64, which cannot be rendered).